feat:予約完了メッセージ

+++ app/Http/Controllers/Admin/UserManagementController.php
@@ -28,13 +28,19 @@ class UserManagementController extends Controller
      */
     public function destroy(Request $request)
     {   
+        if(isset($request->ids)){
 
-        $userIds = $request->ids;
-        foreach ($userIds as $id) {
-            $user = User::find($id);
-            $user->delete();
+            $userIds = $request->ids;
+
+            foreach ($userIds as $id) {
+                $user = User::find($id);
+                $user->reservations()->delete(); // userに紐づく予約データを削除
+                $user->delete();
+            }
+            return redirect()->route('admin.users.index');
+
+        } else {
+            return back();
         }
-
-        return redirect()->route('admin.users.index');
     }
 }
